+++ R/tt_iterate.R
@@ -7,18 +7,20 @@
 #' @param tt_input A `tidytargets` object.
 #' @param target_output Character name of the output target. `NULL` uses an
 #'   auto-generated name.
-#' @param user_function A quoted function call to execute per iteration.
+#' @param command An unevaluated expression. `{targets}` tracks dependencies from
+#'   global symbols in this expression (including upstream target names). Mapped
+#'   targets referenced here also set the iteration pattern.
 #' @param user_function_source_path Optional character path to an R script that
-#'   should be sourced in the worker before calling `user_function`. `NULL`
+#'   should be sourced in the worker before evaluating `command`. `NULL`
 #'   sources nothing.
-#' @param ... Named arguments passed as target inputs; use `is_target()` to
-#'   reference upstream targets by name.
+#' @param ... Additional factory arguments such as `format`, `deployment`,
+#'   or `packages`.
 #'
 #' @export
 tt_iterate <- function(
     tt_input,
     target_output = NULL,
-    user_function = NULL,
+    command = NULL,
     user_function_source_path = NULL,
     ...
 ) {
@@ -30,7 +32,7 @@ tt_iterate <- function(
 tt_iterate.default <- function(
     tt_input,
     target_output = NULL,
-    user_function = NULL,
+    command = NULL,
     user_function_source_path = NULL,
     ...
 ) {
@@ -39,19 +41,17 @@ tt_iterate.default <- function(
 
 #' @rdname tt_iterate
 #' @importFrom glue glue
-#' @importFrom magrittr %>%
 #' @importFrom purrr set_names
 #' @export
 tt_iterate.tidytargets <- function(
     tt_input,
     target_output = NULL,
-    user_function = NULL,
+    command = NULL,
     user_function_source_path = NULL,
     ...
 ) {
     
-    # Check for argument consistency
-    check_for_name_value_conflicts(...)
+    command <- substitute(command)
     
     # Target script
     target_script = glue("{tt_input$initialisation$store}.R")
@@ -61,38 +61,13 @@ tt_iterate.tidytargets <- function(
     
     # Append source if any
     write_source(user_function_source_path, target_script)
-      
-    # please, because sometime we set up list target that do not depend on any other ones
-    # if tiers is set to NULL, then the target will not acquire the _<tier> suffix
-    # I HAVE TO MAKE THIS MORE ELEGANT, AND NOT RELY ON tiers ARGUMENT
-    if(tt_input$initialisation$tier |> get_positions() |> length() < 2){
-      iterate_value = "map"
-      tiers_value = NULL
-    }
-      
-    else if(
-      list(...) |> arguments_to_action(tt_input, "tiered") |> length() == 0 &
-      list(...) |> arguments_to_action(tt_input, "tier") |> length() == 0
-    ){
-      iterate_value = "tier"
-      tiers_value = NULL
-    } else {
-      iterate_value = "tiered"
-      tiers_value = tt_input$initialisation$tier |> get_positions()
-    }
 
     tar_append(
       fx = tt_factory |> quote(),
-      tiers = tiers_value ,
       target_output = target_output,
       script = target_script,
-      user_function = user_function,
-      
-      # I HAVE TO IMPROVE the fact that I have to convert to character 
-      # because arguments_to_action is also used in expand_tiered_arguments, which needs a named vector
-      arguments_to_tier = list(...) |> arguments_to_action(tt_input, "tier") |> as.character()  , # This "tier" value is decided for each new target below. Usually just at the beginning of the piepline
-      arguments_already_tiered = list(...) |> arguments_to_action(tt_input, "tiered") |> as.character(), # This "tiered" value is decided for each new target below. Ususally every other list targets.
-      other_arguments_to_map = list(...) |> arguments_to_action(tt_input, c("tiered", "map")) |> as.character(), # This "tiered" value is decided for each new target below. Ususally every other list targets.
+      command = wrap_quote(command),
+      other_arguments_to_map = command_targets(command, tt_input, "map"),
       ...
     )
   
@@ -101,7 +76,7 @@ tt_iterate.tidytargets <- function(
     tt_input |>
       c(
         as.list(environment())[-1] |> 
-          c(list(iterate = iterate_value)) |> 
+          c(list(iterate = "map")) |> 
           list() |> 
           set_names(target_output) 
       ) |>