+++ R/tt_merge.R
@@ -6,16 +6,18 @@
 #'
 #' @param tt_input A `tidytargets` object.
 #' @param target_output Character name of the output target.
-#' @param user_function A quoted function call to execute for the merge.
+#' @param command An unevaluated expression. `{targets}` tracks dependencies from
+#'   global symbols in this expression (including upstream target names).
 #' @param user_function_source_path Optional character path to an R script to
-#'   source in the worker. `NULL` sources nothing.
-#' @param ... Named arguments passed as target inputs.
+#'   source in the worker before evaluating `command`. `NULL` sources nothing.
+#' @param ... Additional factory arguments such as `format`, `deployment`,
+#'   or `packages`.
 #'
 #' @export
 tt_merge <- function(
     tt_input,
     target_output = NULL,
-    user_function = NULL,
+    command = NULL,
     user_function_source_path = NULL,
     ...
 ) {
@@ -27,7 +29,7 @@ tt_merge <- function(
 tt_merge.default <- function(
     tt_input,
     target_output = NULL,
-    user_function = NULL,
+    command = NULL,
     user_function_source_path = NULL,
     ...
 ) {
@@ -36,19 +38,17 @@ tt_merge.default <- function(
 
 #' @rdname tt_merge
 #' @importFrom glue glue
-#' @importFrom magrittr %>%
 #' @importFrom purrr set_names
 #' @export
 tt_merge.tidytargets <- function(
     tt_input,
     target_output = NULL,
-    user_function = NULL,
+    command = NULL,
     user_function_source_path = NULL,
     ...
 ) {
     
-    # Check for argument consistency
-    check_for_name_value_conflicts(...)
+    command <- substitute(command)
     
     # Target script
     target_script = glue("{tt_input$initialisation$store}.R")
@@ -58,44 +58,15 @@ tt_merge.tidytargets <- function(
     
     # Append source if any
     write_source(user_function_source_path, target_script)
-    
-    
-    # If no tiers
-    if(tt_input$initialisation$tier |> get_positions() |> length() < 2)
-      tar_append(
-          fx = tt_factory |> quote(),
-          target_output = target_output,
-          script = target_script,
-          user_function = user_function,
-          ...
-      )
-      
-    else{
-      
-      args = 
-        list(...)  |> 
-        expand_tiered_arguments(
-          tiers = tt_input$initialisation$tier |> get_positions() |> names(), 
-          argument_to_replace = list(...) |> arguments_to_action(tt_input, "tiered") |> names(),
-          tiered_args = list(...) |> arguments_to_action(tt_input, "tiered") |> names()
-        )
-      
-      # this is needed because I cannot use ellipse (...) anymore, I have to use do.call.
-      do.call(tar_append, c(
-        list(
-          fx = tt_factory |> quote() |> quote(),
-          #tiers = tt_input$initialisation$tier |> get_positions(),
-          target_output = t |> substitute(env = list(t = target_output)) ,
-          script = target_script,
-          user_function = u |> quote() |> substitute(env = list(u = user_function))
-        ),
-        args
-      ))
-    }
 
-    
-    
-    
+    tar_append(
+        fx = tt_factory |> quote(),
+        target_output = target_output,
+        script = target_script,
+        command = wrap_quote(command),
+        ...
+    )
+
     # Add pipeline step
     tt_input |>
       c(